move common stream code to streamcore class
