+++ src/playerWorklet.ts
@@ -1,3 +1,3 @@
-const worker = `var u;(function(f){f.state="state",f.buffer="buffer",f.add="add"})(u||(u={}));var c;(function(f){f.feed="feed",f.id="id"})(c||(c={}));var i;(function(f){f.new="new",f.stale="stale"})(i||(i={}));var a=class extends AudioWorkletProcessor{constructor(e){super(e);this.bindFns(),this.sources=[],this.sourceKey={},this.port.onmessage=s=>this.handleMessage(s),this.handlers={[u.state]:this.handleState,[u.buffer]:this.handleBuffer,[u.add]:this.handleAdd}}bindFns(){this.handleMessage=this.handleMessage.bind(this),this.handleState=this.handleState.bind(this),this.handleBuffer=this.handleBuffer.bind(this),this.handleAdd=this.handleAdd.bind(this),this.bufferKey=this.bufferKey.bind(this),this.process=this.process.bind(this),this.onEndProcess=this.onEndProcess.bind(this)}newStreamItem(e){return{id:e,bufferCursor:0,currentBuffer:0,state:!1,totalBuffers:0,bufferState:i.new}}handleMessage(e){this.handlers[e.data.type](e.data)}handleAdd(e){e.type===u.add&&typeof e.id=="string"&&this.sourceKey[e.id]===void 0&&(this.sourceKey[e.id]=e.index,this.sources[e.index]=this.newStreamItem(e.id))}handleState(e){if(e.type!==u.state||typeof e.state!="boolean"||typeof e.id!="string")return;let s=this.sourceKey[e.id];s!==void 0&&(this.sources[s].state=e.state)}handleBuffer(e){if(e.type!==u.buffer||e.buffer?.buffer===void 0||typeof e.id!="string"||typeof e.bufferID!="string")return;let s=this.sourceKey[e.id];s===void 0&&(s=this.sources.length,this.sourceKey[e.id]=s,this.sources.push(this.newStreamItem(e.id)));let r=this.bufferKey(s);this.sources[s][r]={buffer:e.buffer,id:e.bufferID}}bufferKey(e){let s=this.sources[e].totalBuffers;return this.sources[e].totalBuffers+=1,s}onEndProcess(e){e.length>0&&this.port.postMessage(this.emitBufferIDs(e));let s=[];for(let r=0;r<this.sources.length;r+=1)!this.sources[r].state&&this.sources[r].totalBuffers>0&&(this.sources[r]=this.newStreamItem(this.sources[r].id)),this.sources[r].state&&this.sources[r].totalBuffers-this.sources[r].currentBuffer<6&&s.push(this.sources[r].id);this.port.postMessage(this.emitFeedRequest(s))}emitBufferIDs(e){return{idList:e,type:c.id}}emitFeedRequest(e){return{streams:e,type:c.feed}}process(e,s){try{let r=[];for(let h=0;h<this.sources.length;h+=1){let t=this.sources[h];if(!t||!t.state)continue;let d=s[h];if(!d)continue;let n=d[0];if(!t.state||!t.totalBuffers||!t[t.currentBuffer]){n.fill(0);continue}for(let o=0;o<n.length;o+=1){if(!t.state||!t.totalBuffers||!t[t.currentBuffer]){n.fill(0,o);break}t.bufferState===i.new&&(r.push({bufferID:t[t.currentBuffer].id,sourceID:t.id}),t.bufferState=i.stale),n[o]=t[t.currentBuffer].buffer[t.bufferCursor];let l=!1;if(t.bufferCursor>t[t.currentBuffer].buffer.length-2e3&&t[t.currentBuffer+1]){let b=2e3-(t[t.currentBuffer].buffer.length-t.bufferCursor);n[o]+=t[t.currentBuffer+1].buffer[b],l=!0}t.bufferCursor+=1,t.bufferCursor===t[t.currentBuffer].buffer.length&&(delete t[t.currentBuffer],t.bufferCursor=l?2e3:0,t.currentBuffer+=1,t.bufferState=i.new)}}this.onEndProcess(r)}catch(r){console.warn("Audio Worklet Errored:",r)}return!0}};registerProcessor("playerWorklet",a);`
+const worker = `var u;(function(r){r.state="state",r.buffer="buffer",r.add="add"})(u||(u={}));var o;(function(r){r.feed="feed",r.id="id"})(o||(o={}));var i;(function(r){r.new="new",r.stale="stale"})(i||(i={}));var l=class extends AudioWorkletProcessor{constructor(e){super(e);this.bindFns(),this.sources=[],this.sourceKey={},this.playingBuffer=[],this.requestBuffer=[],this.port.onmessage=t=>this.handleMessage(t),this.handlers={[u.state]:this.handleState,[u.buffer]:this.handleBuffer,[u.add]:this.handleAdd}}bindFns(){this.handleMessage=this.handleMessage.bind(this),this.handleState=this.handleState.bind(this),this.handleBuffer=this.handleBuffer.bind(this),this.handleAdd=this.handleAdd.bind(this),this.bufferKey=this.bufferKey.bind(this),this.process=this.process.bind(this),this.onEndProcess=this.onEndProcess.bind(this)}newStreamItem(e){return{id:e,bufferCursor:0,currentBuffer:0,state:!1,requested:!1,totalBuffers:0,bufferState:i.new}}handleMessage(e){this.handlers[e.data.type](e.data)}handleAdd(e){e.type===u.add&&typeof e.id=="string"&&this.sourceKey[e.id]===void 0&&(this.sourceKey[e.id]=e.index,this.sources[e.index]=this.newStreamItem(e.id))}handleState(e){if(e.type!==u.state||typeof e.state!="boolean"||typeof e.id!="string")return;let t=this.sourceKey[e.id];t!==void 0&&(this.sources[t].state=e.state,this.sources[t].requested=!1)}handleBuffer(e){if(e.type!==u.buffer||e.buffer?.buffer===void 0||typeof e.id!="string"||typeof e.bufferID!="string")return;let t=this.sourceKey[e.id];t===void 0&&(t=this.sources.length,this.sourceKey[e.id]=t,this.sources.push(this.newStreamItem(e.id)));let f=this.bufferKey(t);this.sources[t][f]={buffer:e.buffer,id:e.bufferID},this.sources[t].requested=!1}bufferKey(e){let t=this.sources[e].totalBuffers;return this.sources[e].totalBuffers+=1,t}onEndProcess(){this.playingBuffer.length>0&&this.port.postMessage(this.emitBufferIDs(this.playingBuffer)),this.requestBuffer.length=0;for(let e=0;e<this.sources.length;e+=1)!this.sources[e].state&&this.sources[e].totalBuffers>0&&(this.sources[e]=this.newStreamItem(this.sources[e].id)),this.sources[e].state&&!this.sources[e].requested&&this.sources[e].totalBuffers-this.sources[e].currentBuffer<6&&(this.sources[e].requested=!0,this.requestBuffer.push(this.sources[e].id));this.port.postMessage(this.emitFeedRequest(this.requestBuffer))}emitBufferIDs(e){return{idList:e,type:o.id}}emitFeedRequest(e){return{streams:e,type:o.feed}}process(e,t){try{this.playingBuffer.length=0;for(let f=0;f<this.sources.length;f+=1){let s=this.sources[f];if(!s||!s.state)continue;let d=t[f];if(!d)continue;let n=d[0];if(!s.state||!s.totalBuffers||!s[s.currentBuffer]){n.fill(0);continue}for(let h=0;h<n.length;h+=1){if(!s.state||!s.totalBuffers||!s[s.currentBuffer]){n.fill(0,h);break}s.bufferState===i.new&&(this.playingBuffer.push({bufferID:s[s.currentBuffer].id,sourceID:s.id}),s.bufferState=i.stale),n[h]=s[s.currentBuffer].buffer[s.bufferCursor];let c=!1;if(s.bufferCursor>s[s.currentBuffer].buffer.length-2e3&&s[s.currentBuffer+1]){let a=2e3-(s[s.currentBuffer].buffer.length-s.bufferCursor);n[h]+=s[s.currentBuffer+1].buffer[a],c=!0}s.bufferCursor+=1,s.bufferCursor===s[s.currentBuffer].buffer.length&&(delete s[s.currentBuffer],s.bufferCursor=c?2e3:0,s.currentBuffer+=1,s.bufferState=i.new)}}this.onEndProcess()}catch(f){console.warn("Audio Worklet Errored:",f)}return!0}};registerProcessor("playerWorklet",l);`
 
 export default worker
